[난이도 1] Title: 운동 중독 플레이어, Time: 0.06 ms, Memory: 31.29 MB -BaekjoonHub

+++ goormlevel/195683/1. 운동 중독 플레이어/운동 중독 플레이어.java
@@ -4,7 +4,7 @@ class Main {
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 		String input = br.readLine();
 		String[] strarr = input.split(" ");
-		double answer = Integer.parseInt(strarr[0]) * (double) (1 + (double)(Integer.parseInt(strarr[1]) / 30.0));
+		double answer = Integer.parseInt(strarr[0]) * (1.0 + (double)(Integer.parseInt(strarr[1]) / 30.0));
 		System.out.println((int)answer);
 	}
 }
